+++ src/main/java/gregtech/common/metatileentities/multi/electric/MetaTileEntityPyrolyseOven.java
@@ -14,6 +14,7 @@ import gregtech.common.blocks.BlockWireCoil.CoilType;
 import gregtech.common.blocks.MetaBlocks;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.Vec3i;
 
 public class MetaTileEntityPyrolyseOven extends RecipeMapMultiblockController {
@@ -34,8 +35,8 @@ public class MetaTileEntityPyrolyseOven extends RecipeMapMultiblockController {
     }
 
     @Override
-    protected Vec3i getCenterOffset() {
-        return new Vec3i(1, -1, 0);
+    protected Vec3d getCenterOffset() {
+        return new Vec3d(1, -1, 0);
     }
 
     @Override
