+++ src/main/java/gregtech/common/metatileentities/multi/electric/MetaTileEntityVacuumFreezer.java
@@ -13,6 +13,7 @@ import gregtech.common.blocks.BlockMetalCasing.MetalCasingType;
 import gregtech.common.blocks.MetaBlocks;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.Vec3i;
 
 public class MetaTileEntityVacuumFreezer extends RecipeMapMultiblockController {
@@ -33,8 +34,8 @@ public class MetaTileEntityVacuumFreezer extends RecipeMapMultiblockController {
     }
 
     @Override
-    protected Vec3i getCenterOffset() {
-        return new Vec3i(1, -1, 0);
+    protected Vec3d getCenterOffset() {
+        return new Vec3d(1, -1, 0);
     }
 
     @Override
